Change locking strategy.
One the cache is opened and mapped in memory, there is no need to lock
the "directory db" at each call.
Acquire a shared lock before stat-ing the directory so we can be sure
that nobody is writing to the "directory db". Furthermore, upgrade the
lock to an exlusive lock when we are rebuilding the cache because, in
this function, we may write to the "directory db" if we have to convert
some +CONTENTS files to the +MANIFEST format.