pkg_manifest abstraction via the pkg struct.
We can now access the informations stored in the manifest via the
getters of the struct pkg.
Removed `namever' from the cache (discussed on IRC).
The cache rebuild function use the abstraction, so it will be easy to fetch
the data from the cache itself if the manifest are up to date (only read the
manifest on disk when it is new or updated).