Modified manifest format and database schema.
The struct pkg_file no longer has a size attribute.
The struct pkg has now a flatsize attribute, which represents the size
of the uncompressed package. This attribute is used instead of computing
the value with the old size of files. The attribute is computed in pkg
register and stored in the database, and emited in the manifest.

We also emit @file lines into the manifest so we dont have to read
the whole archive to get informations about a package.