Implement the concept of in-flight package registration.
Split up pkgdb_register_pkg() so that it is expected that callers will call
pkgdb_register_finale() when they are done with the package.  This either
commits the in-flight package registration or rolls it back, depending on
the return code passed to pkgdb_register_finale().

Split up pkg_delete() into component functions that perform individual parts
of the deletion, so they can be used elsewhere.

If pkg_add() fails, clean up any files that were installed.  This probably
needs some refinement, since the add may have failed because the package has
files that also belong to another package.