Add user/group manipulation functions
This is not used yet.
while here fix user/group registration in db which wasn't done before.
users can not be set in manifest in 2 forms: sequence or mapping.
sequence is just for single user name, mapping is for username as key
uid/gid string as value.

The default behaviour is currently unchange, more to come there.

gr_util.c should soon be part of libutil will be dropped from here at that time.