Use a single function to retrive pkg informations:
pkg_get2 is a vaargs function which should get a -1 sentinel.
for user convenience pkg_get is a macros forcing the -1.

to use pkg_get you should 1st provide the field requesred then provide
a reference to the variable.
user of the api should be careful of the type of the variable /!\