Do not mark installed packages as build-only deps
When using --packages-build and/or --delete-build-only, portmaster
could be tricked into marking already installed packages as build-only
dependencies (i.e., second class citizens), thus removing them at the
end, or installing them via packages. To fix it, just check if the
package is already installed, and, should it be, mark it as run
dependency (this is happening in the first pass, which means that
there is no chance that a port installs package A as a build-only
dependency, and then another port marks A as run dependency because it
finds it installed).