pkgdb reopenable + better messages
pkgdb_open is now capable of reopen the database to allow for example to load
the remote database on an already open database.

Better userfriendly messages in case of problem while loading remote databases

modify pkg check so that in case the command is run by a user it just print out
a summary of the problem found, if run as root it proposes to fix the problems
found. the remote(s) database(s) are only opened on demand.

The struct pkgdb * should always be initialized to NULL now