Forced reinstallation/upgrade now possible
Add new -f to install and upgrade subcommands, in case of install it will force
the reinstallation of the package(s) specified. in case of upgrade it will force
the reinstallation of the whole set of packages.

A reinstallation conserver the automatic flags as it used to be.

Better output for upgrade and install, it is now able differenciate:
Downgrading, New installation, Reinstallation upgrade.

Normally if everything goes right you should never notice any downgrading :)