Add the beginnings of ARCH_INDEP package support.
Initially, this modifies pkg register to allow automatically detecting
and marking as architecture independent packages that do not install
any ELF object files.  This should be correct in the vast majority of
cases, if not all.  Any cases where this isn't sufficient probably
aren't going to be dectable heuristically and packagers should
probably set an arch for the pkg explicitly.

Add ARCH_INDEP: yes to pkg.conf to enable this processing: defaults to
no change from existing behaviour.

Sets the arch field in the packages table of local.sqlite to (eg.)
freebsd:9:arch-indep for such packages.