Issue #15 Fix the savepoint/release/rollback to code, which doesn't apparently work with bound parameters.  Just use ordinary string-fu to generate the appropriate commands.
This needs testing over a wider range of conditions than I can
generate.  At the moment, the effect of locking is quite crude, in
that if you run into a locked package it pretty much kills anything
else that might have been going to happen.  However, as this will
interact strongly with the upcoming solver work, better to add the
refinements around that.