Issue #406:
Add new -n flag to pkg(1) itself.  This runs various sanity checks and
aborts without any action if the system is not correctly initialized
to use pkg as its package management system.

'pkg -n' without other arguments should be a good test for a system
being pkgng-enabled.

For best results it will be necessary to modify /usr/sbin/pkg to
recognise the -n flag in command lines and suppress the usual
bootstrapping action.