Enhance plugins management in main().
If a command is not found, when falling-back to plugins, display usage
only if no plugin corresponding to the requested command exist.  If such
a plugin exist and returns an error, do not display usage.

While here remove an inappropriate "not reached" comment on a reached
return.

Fixes this behaviour:
~~~
% pkg Hello World && echo OK
OK
%
~~~