Fix pkg-[r]query
This used to work as it fellthrough to the pkg_rdeps() code.

pkg_rdeps() was looping on the data fetched from pkg_deps()
as PKG_LIST_NEXT uses the passed dep object to find the next
object to lookup.

251ccfc31 removed the break, but left the invalid break still
in the pkg_deps() loop.

This has been effectively wrong since 281bca835