Do not increment lock count if we couldn't lock db
If we could not lock db it is not wise to increment lock count.
At the same time for unlocking db we decrement lock count in any
case as it is unlikely pattern that we could lock db, but could
not unlock and it is nevertheless means that db was unlocked
somehow.