Return EPKG_WARN if the annotatation add/modify/delete didn't affect one row.  In principle this could flag for an annotation affecting several rows, but the SQL can only affect 0 or 1 rows.
In effect, you get EPKG_WARN if you try and add an annotation that
already exists, or delete one that doesn't exist.  'pkg annotate' will
register this as an error, but carry on and apply the annotation to
all requested packages rather than aborting.

Modify would have to go wrong in a weird way to return EPKG_WARN, so
'pkg annotate' treats EPKG_WARN there eqivalently to EPKG_FATAL.