Fix @cwd command handling in plist.
A line that was obtained by getline(3) call should not be considered
as permament storage, hence it is not a good idea to assign it to some
const char pointer. To fix this issue, convert prefix to static array
and copy getlined string to this static array to prevent package
corruption during plist parsing.