Allow a package to only track a repository
Rework the API of annotations so that it only takes an origin.
Add a new pkg_annotation_lookup function
pkgdb_add_annotation does not need to be run in a transaction, which
also allows to reuse it in in the upgrade process.
Restore the repository annotation after an upgrade

To make sure a package is tracking a given repository:
pkg annotate -A mypkg repository <nameoftherepository>