Change manual page so the -d and -r are more mnemonic
-d  Display the list of packages on which pkg-name **depends**.
-r  Display the list of packages which **require** pkg-name.

I don't know if aligning text on the manual page entries for depends and requires to be mnemonic for -d -r is enough. I am using pkg-1.1.1 - in previous releases an informational message "<pkgname> is required by" or "<pkgname> is depends on" accompanied the output when these options were used.  I think the message is important to avoid ambiguity (-d could be "depends" or "dependencies" ; -r could be "requires" or "requirements"), whereas the mnemonics help users establish cached copies of manual pages in their brains.