Fix crash with 'pkg (upgrade|install|upgrade) -r REPO'
The job->reponame is passed into pkgdb_rquery() which later
calls pkg_repo_name() on it. So there is no need to call it
before setting in job->reponame. This was causing it to lookup
again, expecting it to be the ident, but passing in repo-NAME
and resulting in no repository found and then a crash.

Reported by:	David Naylor (dbn@)