Add fallback to yaml parsing.
First try to load pkg.conf via ucl, if failed fallback on yaml. if ok,
Validate the entry against the formats that can be parsed differently in
yaml and ucl, if one of them is not the expected type, then fallback on
yaml.

If the parsing has fallback on yaml, then notify the user and prompt him
the converted version of his pkg.conf file