Hide average speed from the progress meter
Showing two speeds is very confusing. While the average speed can be
interesting at the end of the download, when downloading it's the
current speed which helps identify peaks and stalls. This patch shows
the current speed only, and replaces it with the average when
finished.
As a side point, this leaves more room for package file name.