Update the repo.conf section -- the ``MULTIPLE REPOSITORIES'' concept is now the standard and only way to configure repositories, so don't describe it as a special case.
Move chunks of the description of repo.conf items out of the (doubly
indented) definition list, as the text doesn't fit well in the narrow
column available.

Add discussion of how the order of conf files on REPOS_DIR affects
operation and add examples to match.  Add description of how to
bootstrap pkg(8) from a private repository.

Fixes #842