Split Installation status messages so there is a separate Extract phase.
This moves the progress bar from a single 'Installing/Upgrading' output
to the Extraction message. The old way was showing output from deinstalls
before showing 'Installing/Upgrading and also messages after which was
confused with the next package being installed.

The progress bar was also only capturing % of the extraction phase anyhow
and was missing accounting for all of the other actions of installation
such as scripts. It also was not showing until after the installation
had actually been processed some due to the first tick not being until
extraction.

This removes the need for the meta package extract hack in
pkg_add_common() and do_extract().

  # pkg install -f git
  ...
  [1/1] Reinstalling git-2.0.2_1...
  Updating /etc/shells
  ==> You should manually remove the "git_daemon" user.
  ===> Creating users and/or groups.
  Using existing group 'git_daemon'.
  Using existing user 'git_daemon'.
  [1/1] Extracting git-2.0.2_1: 100%
  Updating /etc/shells

  # pkg install git apache24 xorg
  ...
  [1/3] Installing git-2.0.2_1...
  ===> Creating users and/or groups.
  Using existing group 'git_daemon'.
  Using existing user 'git_daemon'.
  [1/3] Extracting git-2.0.2_1: 100%
  Updating /etc/shells
  [2/3] Installing apache24-2.4.10_1...
  ===> Creating users and/or groups.
  Using existing group 'www'.
  Using existing user 'www'.
  [2/3] Extracting apache24-2.4.10_1: 100%
  /usr/local/share/examples/apache24/httpd.conf -> /usr/local/etc/apache24/httpd.conf
  [3/3] Installing xorg-7.7...

Issue:	#758
Issue:	#923