Don't attempt to set file ownership when INSTALL_AS_USER is set
In this mode we are not running as root so attempting to set file
ownership when installing will most likely fail.

In my tests it didn't actually fail but generated ARCHIVE_WARN,
which pkg interprets as failure.  So one alternative would be
it ignore ARCHIVE_WARN return codes when INSTALL_AS_USER is set.