Implement *printf_pref, pkg_is_jailed and pkg_get_jailname
Messages in pkg were inconsistently prefixed with the name of the jail that
they are running in. There existed no consistent way to ensure that a jail
prefix would be printed out prior to messages from pkg.

This adds a printf_pref family of functions that fill this gap, paving the way
for consistently-prefixed messages from pkg.

To make this work, I've also added helper functions to determine if pkg is
jailed, as well as to get the name of the current jail