Refactor the pkg extraction
In the new process the extraction is more atomic: extract everything
through temporary location and only once the entire package has been
extracted the final rename and credentials settings on the files is
made.

It also simplifies the handling of the rootdir and relocations.

From this point it should be in theory possible to add a rollback in
case of failures per package (this will be done in futur commits)

The extraction now validates that only files and directories described
in the manifest are actually extracted

Reduce a bit the contention on disks if possible in case a file is
already installed by not extracting (still settings the potentially new
credentials)it if the checksums is the same as the new file.