Disable ELF hint checking on NetBSD for now
The ELF hints on NetBSD are optionally located at /var/run/ld.so.hints.
If this file doesn't exist, pkg(8) fails (it should check for it rather
than assume it's there).
However, when this file is touched to create a blank version, pkg(8)
segfaults.  Until support for ELF hints is confirmed to work as it does
on FreeBSD and DragonFly, disable the functionality.