Solve segfault on fetching for NetBSD
The funopen and vasprintf functions are filtered out during <stdio.h>
inclusion on NetBSD unless the _NETBSD_SOURCE macro is defined earlier.

This solves a preplexing segfault on NetBSD.  The funopen function would
be implicitly defined due to the lack of a prototype and the return type
default to an integer, resulting in an invalid pointer to the FILE
structure.