Fix a misleading statement when updating.
When unable to update from the repo the output would look something
like this:

Unable to update repository lab
All repositories are up-to-date.
pkg: Repository lab cannot be opened. 'pkg update' required
Checking for upgrades (0 candidates): 100%
Processing candidates (0 candidates): 100%
Checking integrity... done (0 conflicting)
Your packages are up to date.

and then pkg would exit with a code of 0.

This patch addresses when updating fails and makes it fatal with a
proper exit code.