pkg/unchecked_malloc: update to match code changes
The SmPL expressions were failing to match a lot of cases.  This was due to a
few things:

* Not relaxing function argument checks for malloc/realloc/strdup, etc.
* Not having an optional match terminator at the end of the rule.

This isn't a perfect change; the addition of calling pkg_emit_errno() to
always then return EPKG_FATAL isn't going to work for functions that don't
return an int.  However, this is better than not including anything.