Another follow-up for issue #1431
The initial problem started with commit 4d53b1 which added the
AT_SYMLINK_NOFOLLOW flag to fchmodat function.  On Linux, the flag is
recognized but not implemented.  Setting the flag on Linux causes
ENOTSUP to return in every case, which issue #1431 reported.

The solution to add a second "fallback" attempt by calling fchmodat was
accepted, but I think it was not tested on a port that contained
symbolic links in the packing list.  It doesn't work for symbolic links.
fchmodat will follow the link and fail with ENOENT (file not found).
From my limited research, it's not possible to change file modes on
symbolic links on Linux -- not until AT_SYMLINK_NOFOLLOW is implemented.
Note that lchmod is not an option because it has not been implemented
on Linux, nor is it part of the POSIX standard.

Thus this modification which is only seen by platforms that don't
support AT_SYMLINK_NOFOLLOW (which excludes FreeBSD, DragonFly and
NetBSD but includes Linux) now skips symbolic links as the attempt to
change file mode is impossible and fatal.  This is okay because the
permissions of symlinks have never been evaluated; they are taken from
the target of the link.  Thus changing permissions of symlinks is
cosmetic and the impact of not modifying them is nil in all cases.