Fix progression display
When displaying progression, nbdone contains the number of actions to
action finished, and nbtodl and nbtodl contains respectively the number
of packages to download and the number of packages to process.

Because nbtodl and nbactions can have different values, we should check
if we are downloading (progression is nbdone/nbtodl) or not (in which
case, the progression is nbdone/nbactions).

While here, fix wrapping: when all packages have been downloaded, nbdone
is reset to 0, but when processing packages, this condition can happen
again.  So also reset nbtodl to 0 to avoid this condition.