Rename some variables in security periodic scripts.
Because 410.pkg-audit.in and 460.pkg-checksum.in are not dairy
periodic scripts but security periodic scripts, rename some variables
in them from "dairy_status_security_*" to "security_status_*" to match
them with security periodic scripts in base system. Old
"dairy_status_security_*" variables are still available for backword
compatibility, but waring messages are displayed if they are
explicitly set in /etc/periodic.conf.