Change #include <capability.h> to capsicum.h.
As mentioned in issue #1640, Capsicum has moved from using a header file
called `capability.h` (which could be confused with a Linux header
relating to POSIX.1e "capabilities") to the clearer `capsicum.h`.
Change all instances of `#include <sys/capability.h>` in pkg to
`#include <sys/capsicum.h>`.

This should cause no functional change on systems that have both the old
and the new header file (which includes both 11-STABLE and 10-STABLE),
and it fixes breakage on systems that only have the new header file.
Systems that only have the old header file (e.g., 9-STABLE) will be left
behind: there shouldn't be any build breakage but pkg will no longer be
sandboxed. On the other hand, Capsicum was marked as an experimental
feature on 9-STABLE and that branch has been EOL'ed, so we probably
don't need to worry too much about it.