pkg_elf: add Elf_Note type
On debian if the libelf-dev package is installed elf.h ist taken from
elfutils. Since there is no Elf_Note in the elfutils elf.h compilation
fails in this case.
Fix it by testing for Elf_Note during configuration and typedef Elf_Note
in case the type is not defined.