implement uncompressed meta.conf
pkg can chose the compression format based on the meta configuraiton.
Everything does respect this except the fact the meta itself is
hardcoded to be fetch as .txz (chicken/egg problem).

Given the small size of the meta, lets now authorize to use a meta.conf
at the root of the repo and use it directly, still falling back on the
compressed meta file if not found.

This issue was found by dillon@ (dragonfly) and a solution was discussed
with him.

This commit only concern the consumer of the meta for now, pkg repo
command will be fixed soon