Patch: describe how to change quarterly to latest
Patch for pkg.conf.5 that describes how to change quarterly to latest

I was looking for information about "quarterly" vs "latest" and found the
documentation lacking this part.

The information belongs in the man page for pkg.conf as it contains all the
relevant other information, with the only exception in the details for
"quarterly" and "latest".