Finish 'hash' mode for pkg create
Add two new configuration knobs to control creating packages with
the short hash appended to the package name.

PKG_CREATE_HASH renames the packages to use the short hash
PKG_CREATE_SYMLINK creates a symlink from the original pkg filename

This builds upon b9f1e054731d3dab452cc87c628928033713ee2b and adds
the missing documentation

Also:
pkg create should write to .pkgname first then rename on success

This allows the ports tree to have pkg create output directly to the
destination directory, instead of copying the file afterwards

The hash is separated from the version by a ~ since this character
is not already used as part of a version numbers. Many places in the
code assume the version is everything after the last - character.

Update the fetch code to also use ~ to separate the hash in its cache