Fix #1566: Add pkg-create(8) -l,--level to set compression level
Add -l, --level option to pkg-create(8) to allow specifying a numeric
compression level to be used with one of the compressing -f,--formats, such as
txz, tzst, or tgz.  (If -f tar is specified, along with a level, this is
considered a usage error.)

This expands the libpkg ABI slightly with the new function
pkg_create_set_compression_level().

Additionally, a member in struct pkg_create was changed to one of a different
size; however, (1) struct pkg_create is carefully not part of the ABI, and (2)
the space was likely consumed by padding bytes anyway, so this change should
not break consumers.

The CLI has two special non-numeric compression levels: fast, and best.  They
do what you would expect.

The API has three sentinel compression levels: INT_MIN (==fast), 0 (default),
and INT_MAX (=="best").

There is a defaults change here: if '-f tzst' is specified, we no longer
default to level 20.  Level 20 is extremely slow and the choice is dissimilar
to all other compression formats.  If something desires zstd level 20, it can
invoke pkg-create(8) --format tzst --level 20, or use the corresponding
libpkg(3) API.