Fix child process handling in scripts.c and lua_scripts.c
The WNOWAIT flag is a flag for waitid() and not waitpid(). This results in
an EINVAL return value on macOS and an infinite loop.
Add a new local variable should_waitpid instead to check if the return
status has already been captured.
After this change only two tests fail on macOS and it may also help for
other operating systems.