Add RISC-V ABI aliases
This adds recognition of EM_RISCV ELF files, and aliases for the
following common ABIs:

 - riscv64   (lp64d)
 - riscv64sf (lp64)
 - riscv32   (ilp32d)
 - riscv32sf (ilp32)

On FreeBSD we support riscv64 and its soft-float variant, riscv64sf.
There are no plans to support 32-bit RISC-V at this time, but if we did
this is how I would elect to name it.

This does not attempt to solve the problem of how to name and distribute
packages built with different RISC-V ISA extensions. For now it only
groups based on floating point ABI and ISA width.

The e_flags flags used to determine the float ABI are described in the
RISC-V ELF psABI spec:
https://github.com/riscv/riscv-elf-psabi-doc/blob/master/riscv-elf.md#-file-header