libpkg: correctly re-fetch meta.conf if it's missing
pkg_repo_binary_update currently passes along the timestamp of the database
if the meta.conf is missing, but this is slightly incorrect: if the
meta.conf is missing but we pass in the timestamp of the database, we may
get an incorrect EPKG_UPTODATE because we do not realize that work does need
to be done.

Fix it by moving the got_meta check up one if() to avoid the stat call if we
don't have the .meta, forcing the refetch.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>