libpkg: factor out pkg_jobs_installed_local_pkg()
pkg_jobs_find_remote_pattern used this bit to do a search of the database
when an arbitrary file is specified for the upgrade for confirmation that a
version of it was previously installed.

It will be used to ensure that wildcard matches are discarded if the local
package isn't installed on `pkg upgrade`.