libpkg: only upgrade installed packages with pattern matches
pkg-upgrade provides -g and -x options to upgrade packages matching a
specific pattern.  Currently, it will attempt to install packages that are
not in the dependency chain of a package being upgraded because we only
check that any package matching the query is installed back in
pkg_jobs_find_remote_pattern(), then pulle verything that matches the
pattern.

The first change here is to only do the up-front validation of installation
if we have an exact match.  e.g., if one attempts to `pkg upgrade vim`, we
should check.  If one attempts to `pkg upgrade -g 'FreeBSD-kernel-*'` then
there is no sense in doing this because we end up providing a slightly funky
message:

pkg: FreeBSD-kernel-* is not installed, therefore upgrade is impossible

The second change skips over uninstalled packages if we've used a
glob/regex, so that `pkg upgrade -g 'FreeBSD-kernel-*'` does not pick up
all of the other custom kernels in a given pkgbase repo, for instance.

Fixes #1881