.gitignore: simplify tests/frontend rules
Just assume that everything in tests/frontend should be ignored, unless it's
a *.sh or *.in file, with exception to test_environment.sh which is
generated from test_environment.sh.in.

This reduces the maintenance effort while still making it pretty clear why
a file in this directory is being ignored if it shouldn't be.