message: really expand variables while parsing
for historical reason +DISPLAY was parsed first to ensure it takes
precedence over any message in the +MANIFEST to avoid the message
to be overriden later.

Since the messages has been switch to ucl files, all messages are
appended so there is no needed anymore for this backward compat.

Moreover it was preventing the variable expansions to happen because
the pkg structure was not yet defined