Make -e option orthogonal to other filter options
The -e option had been implemented as a mutual alternative to other
options that select the packages to operate on.

This patch allows to combine -e with other options, the result will be
the intersection of the results of either filter expression.

E.g. the command pkg query -e "%p==${LOCALBASE}" -g %o "py*" will
print the origins of packages whose name starts with "py" and which
have a prefix of $LOCALBASE.

This allows to significantly simplify the CONFLICTS check logic in
bsd.ports.mk, for example.

There are no regressions reported by "make check". All 188 currently
provided test cases succeed. Some testing has been performed to verify
the new functionality, but test cases for the new functionality are
not included in this commit.