Return exit status 1 for additional cases
The combination of -e and package name patterns was not possible
before, therefore there is no precedent whether to return 0 or 1
if there is no match. If -e alone was used, the exit status was 0
even if there was no match, and this behavior is retained.

For the new situation that the package names matched but the
condition failed for all matches, 1 is now returned to indicate
that no output is generated in a query for specific packages.