De-prioritize upgrade-removals as well as deletions
It may be that to handle conflicts we split an upgrade job into separate
removal and installation jobs.  These jobs have special types, but when
sorting we probably want to treat upgrade-removal jobs the same as
deletion jobs.  Otherwise it is possible for the removal job to be
scheduled later than the installation job, leaving the package
effectively uninstalled following an upgrade.