libpkg: Eliminate an unneeded flag variable in struct pkg_solved
We will only split a job of type PKG_SOLVED_UPGRADE.  When we do, it
splits into PKG_SOLVED_UPGRADE_INSTALL and PKG_SOLVED_UPGRADE_REMOVE
jobs, so we will never try to split an already split job.  So the
"already_deleted" flag accomplishes nothing, since it is always false.
Remove it.  Fix a bit of style while here.