libpkg: don't recurse on pkg_jobs_universe_process_deps for local pkgs
This can be reproduced by simply trying to `pkg install` a .pkg out of
/var/cache/pkg.  libpkg ends up blowing up the stack here and crashing,
because it wants to fetch the local package but it's already been presented
with the local package.

Simply stop doing that. This now attempts to install the latest version of
the named packaged from the remote.