pkg repo: rework the balance between the workers
pkg repo works a child CPUs pour create the repository, when doing that
it used to split the list of packages in equal number per worker, which
makes some worker finish before others.

Rework to each worker tell the main process when it is ready and the
main process sends to the work the list of the next package in the
queue.

on a 8 core machines, results:
before the change:
22.76 real        99.02 user         4.57 sys
after the change:
16.69 real       114.12 user         4.52 sys