Add regression tests for pkg-fetch.
This has two tests, one for an unknown package (not in the repo),
another for a package which has a dependency on an unknown package.  It
might be nice to add tests for the case where a package is listed in
the repo, but the package file is absent from the source location.