Add orphans alias to pkg.conf.sample
The orphans alias shows installed packages that are not/no longer
in the pkg repo. This find can help finding outdated installed
packages, that have been renamed, removed from the package list,
or were dropped from the repo for other reasons, or by accident.

In a second step this should probably be added to the pkg periodic
scripts.